# docs/swagger.yaml
basePath: /api
definitions:
  internal_handler.ActivityProductRequest:
    properties:
      activity_price:
        type: number
      activity_stock:
        example: 0
        type: integer
      enable_coupon:
        example: 1
        type: integer
      enable_group_buy:
        example: 0
        type: integer
      group_buy_allow_repeat:
        example: 0
        type: integer
      group_buy_max_joins_per_user:
        example: 1
        type: integer
      group_buy_price:
        type: number
      group_buy_target_count:
        type: integer
      per_user_max_orders:
        example: 0
        type: integer
      per_user_max_qty:
        example: 0
        type: integer
      product_id:
        type: integer
      sort_order:
        type: integer
      status:
        example: 1
        type: integer
    required:
    - activity_price
    - product_id
    type: object
  internal_handler.ActivityRequest:
    properties:
      banner_images:
        items:
          type: string
        type: array
      cover_url:
        type: string
      description:
        type: string
      enable_coupon:
        example: 1
        type: integer
      end_at:
        type: string
      name:
        type: string
      sort_order:
        type: integer
      start_at:
        type: string
      status:
        example: 2
        type: integer
    required:
    - end_at
    - name
    - start_at
    type: object
...
  yujixinjiang_backend_internal_model.Activity:
    properties:
      banner_images:
        items:
          type: string
        type: array
      cover_url:
        type: string
      created_at:
        type: string
      description:
        type: string
      enable_coupon:
        type: integer
      end_at:
        type: string
      id:
        type: integer
      merchant_id:
        type: integer
      name:
        type: string
      products:
        items:
          $ref: '#/definitions/yujixinjiang_backend_internal_model.ActivityProduct'
        type: array
      sort_order:
        type: integer
      start_at:
        type: string
      status:
        type: integer
      updated_at:
        type: string
    type: object
  yujixinjiang_backend_internal_model.ActivityProduct:
    properties:
      activity_id:
        type: integer
      activity_price:
        type: number
      activity_stock:
        type: integer
      created_at:
        type: string
      enable_coupon:
        type: integer
      enable_group_buy:
        type: integer
      group_buy_allow_repeat:
        type: integer
      group_buy_max_joins_per_user:
        type: integer
      group_buy_price:
        type: number
      group_buy_target_count:
        type: integer
      id:
        type: integer
      per_user_max_orders:
        type: integer
      per_user_max_qty:
        type: integer
      product:
        $ref: '#/definitions/yujixinjiang_backend_internal_model.Product'
      product_id:
        type: integer
      sold_count:
        type: integer
      sort_order:
        type: integer
      status:
        type: integer
      updated_at:
        type: string
    type: object
...
  yujixinjiang_backend_internal_model.MerchantProfile:
    properties:
      account:
        $ref: '#/definitions/yujixinjiang_backend_internal_model.Account'
      account_id:
        type: integer
      address:
        type: string
      contact_phone:
        type: string
      created_at:
        type: string
      id:
        type: integer
      images:
        items:
          type: string
        type: array
      shop_logo:
        type: string
      shop_name:
        type: string
      status:
        type: integer
      updated_at:
        type: string
    type: object
...
  yujixinjiang_backend_internal_model.Product:
    properties:
      category:
        $ref: '#/definitions/yujixinjiang_backend_internal_model.ProductCategory'
      category_id:
        type: integer
      cover_url:
        type: string
      created_at:
        type: string
      description:
        type: string
      enable_coupon:
        type: integer
      enable_group_buy:
        type: integer
      group_buy_allow_repeat:
        type: integer
      group_buy_price:
        type: number
      group_buy_target_count:
        type: integer
      id:
        type: integer
      images:
        items:
          type: string
        type: array
      is_hot:
        type: integer
      item_type:
        type: integer
      merchant:
        $ref: '#/definitions/yujixinjiang_backend_internal_model.MerchantProfile'
      merchant_id:
        type: integer
      name:
        type: string
      original_price:
        type: number
      price:
        type: number
      sales_count:
        type: integer
      status:
        type: integer
      stock:
        type: integer
      updated_at:
        type: string
    type: object
  yujixinjiang_backend_internal_model.ProductCategory:
    properties:
      created_at:
        type: string
      icon_url:
        type: string
      id:
        type: integer
      merchant_id:
        type: integer
      name:
        type: string
      parent_id:
        type: integer
      sort_order:
        type: integer
      status:
        type: integer
      updated_at:
        type: string
    type: object
...
  yujixinjiang_backend_internal_query.PageResult:
    properties:
      list: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  yujixinjiang_backend_internal_response.Body:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  yujixinjiang_backend_internal_service.ActivityStoreView:
    properties:
      banner_images:
        items:
          type: string
        type: array
      cover_url:
        type: string
      created_at:
        type: string
      description:
        type: string
      enable_coupon:
        type: integer
      end_at:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      merchant_id:
        type: integer
      name:
        type: string
      products:
        items:
          $ref: '#/definitions/yujixinjiang_backend_internal_model.ActivityProduct'
        type: array
      sort_order:
        type: integer
      start_at:
        type: string
      status:
        type: integer
      updated_at:
        type: string
    type: object
...
info:
  contact:
    email: support@yujixinjiang.com
    name: API Support
  description: 豫记信疆微信小程序后端接口文档，支持在线调试。
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: 豫记信疆 API
  version: "1.0"
paths:
  /activities/{id}:
    get:
      parameters:
      - description: 活动 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/yujixinjiang_backend_internal_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/yujixinjiang_backend_internal_service.ActivityStoreView'
              type: object
      summary: 活动详情（子商店）
      tags:
      - 用户-商城
  /activities/{id}/products:
    get:
      parameters:
      - description: 活动 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/yujixinjiang_backend_internal_response.Body'
      summary: 活动商品列表（子商店）
      tags:
      - 用户-商城
...
  /merchant/activities:
    get:
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页条数
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/yujixinjiang_backend_internal_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/yujixinjiang_backend_internal_query.PageResult'
              type: object
      security:
      - BearerAuth: []
      summary: 本店活动列表
      tags:
      - 商家端-活动
    post:
      consumes:
      - application/json
      parameters:
      - description: 活动信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.ActivityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/yujixinjiang_backend_internal_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/yujixinjiang_backend_internal_model.Activity'
              type: object
      security:
      - BearerAuth: []
      summary: 创建活动
      tags:
      - 商家端-活动
  /merchant/activities/{id}:
    delete:
      parameters:
      - description: 活动 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/yujixinjiang_backend_internal_response.Body'
      security:
      - BearerAuth: []
      summary: 删除活动
      tags:
      - 商家端-活动
    get:
      parameters:
      - description: 活动 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/yujixinjiang_backend_internal_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/yujixinjiang_backend_internal_model.Activity'
              type: object
      security:
      - BearerAuth: []
      summary: 活动详情（含商品）
      tags:
      - 商家端-活动
    patch:
      consumes:
      - application/json
      parameters:
      - description: 活动 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 活动信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.ActivityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/yujixinjiang_backend_internal_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/yujixinjiang_backend_internal_model.Activity'
              type: object
      security:
      - BearerAuth: []
      summary: 更新活动
      tags:
      - 商家端-活动
  /merchant/activities/{id}/products:
    post:
      consumes:
      - application/json
      parameters:
      - description: 活动 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 活动商品配置
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.ActivityProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/yujixinjiang_backend_internal_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/yujixinjiang_backend_internal_model.ActivityProduct'
              type: object
      security:
      - BearerAuth: []
      summary: 添加活动商品
      tags:
      - 商家端-活动
  /merchant/activities/{id}/products/{product_id}:
    delete:
      parameters:
      - description: 活动 ID
        in: path
        name: id
        required: true
        type: integer
      - description: activity_product.id
        in: path
        name: product_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/yujixinjiang_backend_internal_response.Body'
      security:
      - BearerAuth: []
      summary: 移除活动商品
      tags:
      - 商家端-活动
    patch:
      consumes:
      - application/json
      parameters:
      - description: 活动 ID
        in: path
        name: id
        required: true
        type: integer
      - description: activity_product.id
        in: path
        name: product_id
        required: true
        type: integer
      - description: 活动商品配置
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.ActivityProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/yujixinjiang_backend_internal_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/yujixinjiang_backend_internal_model.ActivityProduct'
              type: object
      security:
      - BearerAuth: []
      summary: 更新活动商品
      tags:
      - 商家端-活动
...
  /merchants/{id}/activities:
    get:
      parameters:
      - description: 商家 ID
        in: path
        name: merchant_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/yujixinjiang_backend_internal_response.Body'
      summary: 商家进行中的活动列表
      tags:
      - 用户-商城
...
securityDefinitions:
  BearerAuth:
    description: '登录后获取的 JWT，格式: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"